# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  code?: string;
}

export type DataStatusState = typeof DataStatusState[keyof typeof DataStatusState];


export const DataStatusState = {
  live: 'live',
  demo: 'demo',
  offline: 'offline',
  partial: 'partial',
} as const;

export interface DataStatus {
  state: DataStatusState;
  provider: string;
  /** @nullable */
  lastUpdated: string | null;
  message: string;
}

export interface ModelPerformance {
  market: string;
  sampleSize: number;
  /** @nullable */
  hitRate: number | null;
  /** @nullable */
  brierScore: number | null;
  /** @nullable */
  mae: number | null;
  /** @nullable */
  rmse: number | null;
  status: string;
}

export interface DashboardSummary {
  dataStatus: DataStatus;
  games: number;
  snipes: number;
  playersAnalyzed: number;
  /** @nullable */
  topConfidence: string | null;
  modelVersion: string;
  marketPerformance: ModelPerformance[];
}

export interface Team {
  id: string;
  name: string;
  /** @nullable */
  city: string | null;
  abbreviation: string;
  /** @nullable */
  conference: string | null;
  /** @nullable */
  division: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  primaryColor: string | null;
  /** @nullable */
  secondaryColor: string | null;
}

export interface Player {
  id: string;
  fullName: string;
  firstName: string;
  lastName: string;
  teamId: string;
  team: Team;
  position: string;
  /** @nullable */
  jerseyNumber: number | null;
  /** @nullable */
  headshotUrl: string | null;
}

export interface TeamProfile {
  team: Team;
  /** @nullable */
  offense: number | null;
  /** @nullable */
  defense: number | null;
  /** @nullable */
  goaltending: number | null;
  /** @nullable */
  powerPlay: number | null;
  /** @nullable */
  penaltyKill: number | null;
  /** @nullable */
  shotGeneration: number | null;
  /** @nullable */
  shotSuppression: number | null;
  /** @nullable */
  expectedGoals: number | null;
}

export type TeamDetail = Team & {
  roster: Player[];
  profile: TeamProfile;
};

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  scheduled: 'scheduled',
  live: 'live',
  final: 'final',
  postponed: 'postponed',
  unknown: 'unknown',
} as const;

export interface Game {
  id: string;
  gameDate: string;
  awayTeam: Team;
  homeTeam: Team;
  /** @nullable */
  venue: string | null;
  status: GameStatus;
  /** @nullable */
  statusDetail: string | null;
  /** @nullable */
  matchupScore: number | null;
  /** @nullable */
  goalEnvironment: number | null;
  /** @nullable */
  shotEnvironment: number | null;
  /** @nullable */
  powerPlayEdge: number | null;
  /** @nullable */
  goaltendingEdge: number | null;
}

export type GameDetail = Game & {
  notes: string[];
};

export interface StatLine {
  label: string;
  games: number;
  goals: number;
  sog: number;
  points: number;
  assists: number;
  toi: number;
  ppToi: number;
}

export type PlayerDetail = Player & {
  seasonStats: StatLine;
  recentStats: StatLine[];
  splits: StatLine[];
};

export type GoalieStatus = typeof GoalieStatus[keyof typeof GoalieStatus];


export const GoalieStatus = {
  confirmed: 'confirmed',
  projected: 'projected',
  unknown: 'unknown',
} as const;

export interface Goalie {
  id: string;
  fullName: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  jerseyNumber: number | null;
  /** @nullable */
  headshotUrl: string | null;
  team: Team;
  opponent: Team | null;
  status: GoalieStatus;
  /** @nullable */
  projectedSaves: number | null;
  /** @nullable */
  projectedShotsAgainst: number | null;
  /** @nullable */
  savePercentage: number | null;
  /** @nullable */
  goalsAgainst: number | null;
  /** @nullable */
  winProbability: number | null;
  /** @nullable */
  shutoutProbability: number | null;
  /** @nullable */
  saves25Plus: number | null;
  /** @nullable */
  saves30Plus: number | null;
  /** @nullable */
  saves35Plus: number | null;
}

export type PropLineStatus = typeof PropLineStatus[keyof typeof PropLineStatus];


export const PropLineStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface Prop {
  id: string;
  player: Player;
  opponent: Team;
  market: string;
  /** @nullable */
  line: number | null;
  lineStatus: PropLineStatus;
  /** @nullable */
  modelProjection: number | null;
  /** @nullable */
  overProbability: number | null;
  /** @nullable */
  underProbability: number | null;
  /** @nullable */
  edge: number | null;
  /** @nullable */
  confidence: string | null;
  /** @nullable */
  source: string | null;
}

export interface Snipe {
  id: string;
  player: Player;
  opponent: Team;
  score: number;
  /** @nullable */
  goalProbability: number | null;
  /** @nullable */
  sogProjection: number | null;
  /** @nullable */
  pointProbability: number | null;
  /** @nullable */
  assistProbability: number | null;
  /** @nullable */
  modelProjection: number | null;
  /** @nullable */
  line: number | null;
  /** @nullable */
  edge: number | null;
  confidence: string;
  factors: string[];
}

export interface Matchup {
  game: Game;
  away: TeamProfile;
  home: TeamProfile;
}

export type AuditRecordResult = typeof AuditRecordResult[keyof typeof AuditRecordResult];


export const AuditRecordResult = {
  hit: 'hit',
  miss: 'miss',
  pending: 'pending',
  void: 'void',
} as const;

export interface AuditRecord {
  id: string;
  createdAt: string;
  playerName: string;
  market: string;
  /** @nullable */
  line: number | null;
  /** @nullable */
  projection: number | null;
  /** @nullable */
  probability: number | null;
  /** @nullable */
  confidence: string | null;
  modelVersion: string;
  /** @nullable */
  actual: number | null;
  result: AuditRecordResult;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type DateQueryParameter = string;

export type SearchQueryParameter = string;

export type TeamQueryParameter = string;

export type MarketQueryParameter = typeof MarketQueryParameter[keyof typeof MarketQueryParameter];


export const MarketQueryParameter = {
  goals: 'goals',
  sog: 'sog',
  points: 'points',
  assists: 'assists',
  saves: 'saves',
  blocks: 'blocks',
  hits: 'hits',
} as const;

export type GetGamesParams = {
date?: DateQueryParameter;
};

export type GetPlayersParams = {
search?: SearchQueryParameter;
team?: TeamQueryParameter;
};

export type GetPropsParams = {
market?: MarketQueryParameter;
};

export type GetSnipesParams = {
market?: MarketQueryParameter;
};