# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { DataStatus } from './dataStatus';
import type { ModelPerformance } from './modelPerformance';

export interface DashboardSummary {
  dataStatus: DataStatus;
  games: number;
  snipes: number;
  playersAnalyzed: number;
  /** @nullable */
  topConfidence: string | null;
  modelVersion: string;
  marketPerformance: ModelPerformance[];
}