# lib/api-zod/src/generated/types/dataStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { DataStatusState } from './dataStatusState';

export interface DataStatus {
  state: DataStatusState;
  provider: string;
  /** @nullable */
  lastUpdated: string | null;
  message: string;
}