# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';
import type { Team } from './team';

export interface Game {
  id: string;
  gameDate: Date;
  awayTeam: Team;
  homeTeam: Team;
  /** @nullable */
  venue: string | null;
  status: GameStatus;
  /** @nullable */
  statusDetail: string | null;
  /** @nullable */
  matchupScore: number | null;
  /** @nullable */
  goalEnvironment: number | null;
  /** @nullable */
  shotEnvironment: number | null;
  /** @nullable */
  powerPlayEdge: number | null;
  /** @nullable */
  goaltendingEdge: number | null;
}