# lib/api-zod/src/generated/types/goalie.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalieStatus } from './goalieStatus';
import type { Team } from './team';

export interface Goalie {
  id: string;
  fullName: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  jerseyNumber: number | null;
  /** @nullable */
  headshotUrl: string | null;
  team: Team;
  opponent: Team | null;
  status: GoalieStatus;
  /** @nullable */
  projectedSaves: number | null;
  /** @nullable */
  projectedShotsAgainst: number | null;
  /** @nullable */
  savePercentage: number | null;
  /** @nullable */
  goalsAgainst: number | null;
  /** @nullable */
  winProbability: number | null;
  /** @nullable */
  shutoutProbability: number | null;
  /** @nullable */
  saves25Plus: number | null;
  /** @nullable */
  saves30Plus: number | null;
  /** @nullable */
  saves35Plus: number | null;
}