# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Team } from './team';

export interface Player {
  id: string;
  fullName: string;
  firstName: string;
  lastName: string;
  teamId: string;
  team: Team;
  position: string;
  /** @nullable */
  jerseyNumber: number | null;
  /** @nullable */
  headshotUrl: string | null;
}