# lib/api-zod/src/generated/types/playerDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { StatLine } from './statLine';

export type PlayerDetail = Player & {
  seasonStats: StatLine;
  recentStats: StatLine[];
  splits: StatLine[];
};