# lib/api-zod/src/generated/types/prop.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { PropLineStatus } from './propLineStatus';
import type { Team } from './team';

export interface Prop {
  id: string;
  player: Player;
  opponent: Team;
  market: string;
  /** @nullable */
  line: number | null;
  lineStatus: PropLineStatus;
  /** @nullable */
  modelProjection: number | null;
  /** @nullable */
  overProbability: number | null;
  /** @nullable */
  underProbability: number | null;
  /** @nullable */
  edge: number | null;
  /** @nullable */
  confidence: string | null;
  /** @nullable */
  source: string | null;
}