# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: string;
  name: string;
  /** @nullable */
  city: string | null;
  abbreviation: string;
  /** @nullable */
  conference: string | null;
  /** @nullable */
  division: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  primaryColor: string | null;
  /** @nullable */
  secondaryColor: string | null;
}