# lib/api-zod/src/generated/types/teamDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * NHLsnipes analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { Team } from './team';
import type { TeamProfile } from './teamProfile';

export type TeamDetail = Team & {
  roster: Player[];
  profile: TeamProfile;
};